add authorization

+++ main.py
@@ -1,13 +1,19 @@
 from sqlalchemy.exc import IntegrityError
 from fastapi import FastAPI, Depends, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
+from auth import verify_password, create_access_token, SECRET_KEY, ALGORITHM
 from sqlalchemy.orm import Session
 from sqlalchemy import func
 from typing import List
+from jose import jwt, JWTError
 
 import models
 import schemas
+from auth import get_password_hash
 from database import engine, get_db
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
 
 models.Base.metadata.create_all(bind=engine)
 
@@ -21,17 +27,43 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+def get_current_user(token: str = Depends(oauth2_scheme)):
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        user_id: str = payload.get("sub")
+        if user_id is None:
+            raise HTTPException(status_code=401, detail="Невалидный токен")
+        return int(user_id)
+    except (JWTError, ValueError):
+        raise HTTPException(status_code=401, detail="Ошибка авторизации")
+
+@app.post("/login")
+def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
+    user = db.query(models.User).filter(models.User.username == form_data.username).first()
+    if not user or not verify_password(form_data.password, user.hashed_password):
+        raise HTTPException(
+            status_code=401,
+            detail="Неверное имя пользователя или пароль",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token = create_access_token(data={"sub": str(user.id)})
+    return {"access_token": access_token, "token_type": "bearer"}
+
 @app.get("/users", response_model=List[schemas.UserOut])
 def get_users(db: Session = Depends(get_db)):
     return db.query(models.User).all()
 
 @app.post("/users", response_model=schemas.UserOut)
 def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    db_user = models.User(**user.model_dump())
-    db.add(db_user)
+    db_user = db.query(models.User).filter(models.User.username == user.username).first()
+    if db_user:
+        raise HTTPException(status_code=400, detail="Имя пользователя уже занято")
+    hashed_pwd = get_password_hash(user.password)
+    new_user = models.User(username=user.username, email=user.email, hashed_password=hashed_pwd)
+    db.add(new_user)
     db.commit()
-    db.refresh(db_user)
-    return db_user
+    db.refresh(new_user)
+    return new_user
 
 @app.get("/recipes", response_model=List[schemas.RecipeOut])
 def get_recipes(search: str = None, db: Session = Depends(get_db)):
@@ -43,8 +75,8 @@ def get_recipes(search: str = None, db: Session = Depends(get_db)):
     return query.all()
 
 @app.post("/recipes", response_model=schemas.RecipeOut)
-def create_recipe(recipe: schemas.RecipeCreate, db: Session = Depends(get_db)):
-    db_recipe = models.Recipe(**recipe.model_dump())
+def create_recipe(recipe: schemas.RecipeCreate, db: Session = Depends(get_db), current_user_id: int = Depends(get_current_user)):
+    db_recipe = models.Recipe(**recipe.model_dump(), author_id=current_user_id)
     try:
         db.add(db_recipe)
         db.commit()
@@ -55,10 +87,12 @@ def create_recipe(recipe: schemas.RecipeCreate, db: Session = Depends(get_db)):
         raise HTTPException(status_code=400, detail="Рецепт с таким названием уже существует")
 
 @app.delete("/recipes/{recipe_id}")
-def delete_recipe(recipe_id: int, db: Session = Depends(get_db)):
+def delete_recipe(recipe_id: int, db: Session = Depends(get_db), current_user_id: int = Depends(get_current_user)):
     db_recipe = db.query(models.Recipe).get(recipe_id)
     if not db_recipe:
         raise HTTPException(status_code=404, detail="Рецепт не найден")
+    if db_recipe.author_id != current_user_id:
+        raise HTTPException(status_code=403, detail="Не твой рецепт")
     db.delete(db_recipe)
     db.commit()
     return {"message": "Рецепт был удалён"}
